# HV-1566 Cache type variable index for performance reasons