- https://hibernate.atlassian.net/browse/HV-1566

and same results as in the JIRA ticket:
```bash
current master:

# Run complete. Total time: 00:01:52

Benchmark                                                                                   Mode  Cnt    Score    Error   Units
CascadedValidation.testCascadedValidation                                                  thrpt   20  160.152 ± 11.923  ops/ms
CascadedWithLotsOfItemsAndMoreConstraintsValidation.testCascadedValidationWithLotsOfItems  thrpt   20  471.770 ±  8.797   ops/s
CascadedWithLotsOfItemsValidation.testCascadedValidationWithLotsOfItems                    thrpt   20  633.382 ± 11.182   ops/s

with changes

# Run complete. Total time: 00:01:50

Benchmark                                                                                   Mode  Cnt    Score    Error   Units
CascadedValidation.testCascadedValidation                                                  thrpt   20  170.439 ± 15.699  ops/ms
CascadedWithLotsOfItemsAndMoreConstraintsValidation.testCascadedValidationWithLotsOfItems  thrpt   20  471.262 ± 20.874   ops/s
CascadedWithLotsOfItemsValidation.testCascadedValidationWithLotsOfItems                    thrpt   20  703.661 ± 39.131   ops/s
```