@gsmet 2018-02-06
OK, this looks good and it's a good thing to improve the container element case. I don't think we have benchmarked it a lot. Will merge.