@gsmet 2018-02-08
Merged with a few further minor tweaks: in our getter, we usually don't use `this.` (at least not in the modified classes) so I removed them.

Thanks!