@marko-bekhta 2018-02-08
> we usually don't use this.

Ahh, ok. That's just how the IDE generates them. Will fix that for HV settings :)

> wondering if we shouldn't introduce our own structure (TypeParameter?) wrapping the values

That's actually very close to what I have in mind for a different PR I wanted to discuss. I was just waiting while we finish with the opened ones :)